Replace use of deprecated package analysis with a newer one.
Package analysis has been deprecated. It's old, and was created as an
ad-hoc best-effort solution before there was a specification for
declaring a .go file as generated.

It has been superseded by package github.com/shurcooL/go/generated,
which implements that specification.

This change updates the code and README to use the newer package.
The behavior is preserved as much as possible, but it should be more
accurate at detecting the "// Code generated … DO NOT EDIT." comment.